renamed UserViolationsService to ViolationService, moved all methods connected with violations into ViolationService
