+++ core/src/test/java/greencity/controller/ManagementOrderControllerTest.java
@@ -8,6 +8,7 @@ import greencity.filters.CertificateFilterCriteria;
 import greencity.filters.CertificatePage;
 import greencity.service.ubs.CertificateService;
 import greencity.service.ubs.UBSManagementService;
+import greencity.service.ubs.ViolationService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -43,6 +44,9 @@ class ManagementOrderControllerTest {
 
     @Mock
     UBSManagementService ubsManagementService;
+
+    @Mock
+    private ViolationService violationService;
 
     @Mock
     CertificateService certificateService;
@@ -168,18 +172,18 @@ class ManagementOrderControllerTest {
         this.mockMvc.perform(get(ubsLink + "/violation-details" + "/{orderId}", 1L))
             .andExpect(status().isNotFound());
 
-        verify(ubsManagementService).getViolationDetailsByOrderId(1L);
+        verify(violationService).getViolationDetailsByOrderId(1L);
     }
 
     @Test
     void returnsDetailsAboutViolationWithGivenOrderId() throws Exception {
         ViolationDetailInfoDto violationDetailInfoDto = getViolationDetailInfoDto();
-        when(ubsManagementService.getViolationDetailsByOrderId(1L)).thenReturn(Optional.of(violationDetailInfoDto));
+        when(violationService.getViolationDetailsByOrderId(1L)).thenReturn(Optional.of(violationDetailInfoDto));
 
         this.mockMvc.perform(get(ubsLink + "/violation-details" + "/{orderId}", 1L))
             .andExpect(status().isOk());
 
-        verify(ubsManagementService).getViolationDetailsByOrderId(1L);
+        verify(violationService).getViolationDetailsByOrderId(1L);
     }
 
     @Test
@@ -237,7 +241,7 @@ class ManagementOrderControllerTest {
         mockMvc.perform(delete(ubsLink + "/delete-violation-from-order" + "/{orderId}", 1L))
             .andExpect(status().isOk());
 
-        verify(ubsManagementService).deleteViolation(1L, null);
+        verify(violationService).deleteViolation(1L, null);
     }
 
     @Test
