+++ service-api/src/main/java/greencity/service/ubs/UBSManagementService.java
@@ -106,16 +106,6 @@ public interface UBSManagementService {
     ViolationsInfoDto getAllUserViolations(String email);
 
     /**
-     * Method for adding violation for user.
-     *
-     * @param add            {@link AddingViolationsToUserDto}
-     * @param multipartFiles {@link MultipartFile}
-     * @param uuid           {@link String}.
-     * @author Nazar Struk
-     */
-    void addUserViolation(AddingViolationsToUserDto add, MultipartFile[] multipartFiles, String uuid);
-
-    /**
      * Method for send email with description to user.
      *
      * @param dto {@link AddingViolationsToUserDto } order id with description.
@@ -188,15 +178,6 @@ public interface UBSManagementService {
     List<DetailsOrderInfoDto> getOrderBagsDetails(Long orderId);
 
     /**
-     * Method returns detailed information about user violation by order id.
-     *
-     * @param orderId of {@link Long} order id;
-     * @return {@link ViolationDetailInfoDto};
-     * @author Rusanovscaia Nadejda
-     */
-    Optional<ViolationDetailInfoDto> getViolationDetailsByOrderId(Long orderId);
-
-    /**
      * Method that get order and payment status.
      *
      * @author Mahdziak Orest
@@ -230,15 +211,6 @@ public interface UBSManagementService {
      * @author Nazar Struk
      */
     List<AdditionalBagInfoDto> getAdditionalBagsInfo(Long orderId);
-
-    /**
-     * Method deletes violation from database by orderId.
-     *
-     * @param orderId {@link Long}
-     * @param uuid    {@link String}.
-     * @author Nadia Rusanovscaia
-     */
-    void deleteViolation(Long orderId, String uuid);
 
     /**
      * Method that saves manual payment and returns response with required fields.
@@ -288,15 +260,6 @@ public interface UBSManagementService {
      * Method that update EmployeePositionDtoResponse.
      */
     void updatePositions(EmployeePositionDtoResponse dto, String uuid);
-
-    /**
-     * Method for adding violation for user.
-     *
-     * @param add            {@link AddingViolationsToUserDto}
-     * @param multipartFiles {@link MultipartFile}
-     * @author Bohdan Melnyk
-     */
-    void updateUserViolation(UpdateViolationToUserDto add, MultipartFile[] multipartFiles, String uuid);
 
     /**
      * Method that save ReasonNotTakeBagDto.
